requirements.txt
langgraph
groq
pandas
numpy
yfinance
python-dotenv
pytest
